feat: stop enforcing /mcp in urls used with MCPToolbox (#5028)
We were trailing the received URL with `/mcp` when initializing our MCP
toolbox for databases class. It is conventional to expose mcp servers
under `.../mcp`, but not required by the protocol, so we shouldn't
enforce it like this.

Added `append_mcp_to_url: bool = True` to let users opt-out while
keeping backwards compatibility.