Added MCP server prompts functionality
Added new code into the `initialize()` method of the `MCPTools()` class to fetch and list mcp server prompts and make it available to the agent as part of the MCPTools tooklit.
Added a new `get_entrypoint_for_prompt()` function and imported it into MCPTools() class to aid in the above changes.